Restructure components for better UX
