+++ client/components/ArtGallery.js
@@ -1,85 +1,62 @@
 import React from 'react'
-import {ModalContainer, ModalDialog} from 'react-modal-dialog';
-import ReactSpinner from 'react-spinjs';
-import SearchInput, {createFilter} from 'react-search-input'
-import NavBar from './NavBar'
+import AuthModal from './AuthModal'
+import ArtWindow from './ArtWindow'
+import * as auth from '../models/auth'
+import * as art from '../models/art'
 
-const KEYS_TO_FILTERS = ['Artist Name', 'Art Title']
-
-export default class ArtGallery extends React.Component {
+export default class App extends React.Component {
   constructor(props) {
-    super(props)
+    super(props);
 
     this.state = {
-      showInfo: false,
-      searchTerm: ''
+      loggedIn: false,
+      artCollection: []
     }
   }
 
-  parseImageUrl(imgUrl) {
-    imgUrl = imgUrl.split(';')
-    return imgUrl
+  componentWillMount() {
+    this.fetchArt(this.props.params.artistName)
   }
-  openInfo(art) {
-    this.setState({showInfo: true});
-    this.setState({currentArt: art})
+  signUp(userData) {
+    console.log('signing up in~~~app.js')
+    auth.signUp(userData)
+    this.setState({loggedIn: true})
   }
-  closeInfo() {
-    this.setState({showInfo: false});
+  login(userData) {
+    console.log('logging in~~~app.js')
+    auth.login(userData)
+    this.setState({loggedIn: true})
   }
-  searchUpdated (term) {
-    this.setState({searchTerm: term})
+  logout() {
+    console.log('logging out')
+    this.setState({loggedIn: false})
+  }
+  fetchArt(artist) {
+    art.getArt()
+    .then((artwork) => {
+      if(artist) {
+        this.setState({artCollection: artwork.filter((art) => art['Artist Name'] == artist)})
+      }
+      else {
+        this.setState({artCollection: artwork})
+      }   
+    })
   }
 
-  render() {
-    const filteredArt = this.props.gallery.filter(createFilter(this.state.searchTerm, KEYS_TO_FILTERS))
+  render(){
     return (
       <div>
-      <SearchInput className="search-input" onChange={this.searchUpdated.bind(this)} />
-      <div className="artGallery">
-        <NavBar />
-        {this.state.showInfo ?
-          <Info onClose={this.closeInfo.bind(this)} currentArt={this.state.currentArt} parseImageUrl={this.parseImageUrl.bind(this)}/>
-        : null} 
-
-        {filteredArt.map((art) => {
-          return <div className="artwork" key={art._id}>
-            <img className="artImage" src={this.parseImageUrl(art.Images)[0]} onClick={(e) => this.openInfo(art)}/>            
-            {this.props.loggedIn ?
-              <div className="userFeatures">
-              <button>Like</button>
-              <button>Fav!</button>
-              </div> 
-              : ''}
-          </div>
-        })}
-      </div>
+        <h2>Austin Art</h2>
+        
+        <AuthModal className="loginButton" login={this.login.bind(this)} signUp={this.signUp.bind(this)}/>
+        <ArtWindow className="artGallery" gallery={this.state.artCollection} loggedIn={this.state.loggedIn} fetchArt={this.fetchArt.bind(this)}/>
       </div>
     )
   }
 }
 
-class Info extends React.Component {
 
-  render() {
-    return (
-      <ModalContainer onClose={this.props.onClose}>
-     
-        <ModalDialog onClose={this.props.onClose} className="info">
-         
-            <h2>{this.props.currentArt['Art Title']}</h2>
-            <p>By: {this.props.currentArt['Artist Name']}</p>
-   
-            <img src={this.props.parseImageUrl(this.props.currentArt.Images)[0]} />
-            <img src={this.props.parseImageUrl(this.props.currentArt.Images)[1]} />
-            <img src={this.props.parseImageUrl(this.props.currentArt.Images)[2]}/>
-                  
-        </ModalDialog>
-      
-    </ModalContainer>
-    )
-  }
-}
+
