+++ client/public/index.html
@@ -8,7 +8,6 @@
   <link rel="stylesheet" href="http://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.6.3/css/font-awesome.min.css">
   <style>
   body {font-family: "Lato", sans-serif}
-  .mySlides {display: none}
   </style>
   <link rel="stylesheet" href="/styles.css"/>
 </head>
